+++ src/main/java/com/octagram/pollet/survey/infrastructure/persistence/QuestionRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.octagram.pollet.survey.infrastructure;
+package com.octagram.pollet.survey.infrastructure.persistence;
 
 import static com.octagram.pollet.survey.domain.model.QQuestion.*;
 import static com.octagram.pollet.survey.domain.model.QQuestionOption.*;
