+++ src/main/java/com/octagram/pollet/survey/infrastructure/persistence/SurveyTagRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.octagram.pollet.survey.infrastructure;
+package com.octagram.pollet.survey.infrastructure.persistence;
 
 import com.octagram.pollet.survey.domain.model.QSurveyTag;
 import com.octagram.pollet.survey.domain.model.QTag;
